+++ .cargo/config
@@ -1,0 +1,2 @@
+[alias]
+xtask = "run --package xtask --"
