+++ README.md
@@ -8,6 +8,14 @@ This repository houses the non-compiler components of the system.
 
 The compiler repo can be found [here](https://github.com/softdevteam/ykrustc).
 
+## Getting Started
+
+To get started, see
+[this section in the documentation](https://softdevteam.github.io/ykdocs/dev/getting_started.html).
+
+This repo uses xtask instead of regular cargo, so instead of running `cargo
+<target>`, you will instead need to execute `cargo xtask <target>`.
+
 ## Contributors
 
 Yorick is developed by a team of people listed in the
