+++ internal_ws/ykbh/src/lib.rs
@@ -269,112 +269,3 @@ impl SIRInterpreter {
         }
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::SIRInterpreter;
-    use yktrace::sir::SIR;
-
-    fn interp(fname: &str, tio: *mut u8) {
-        let body = SIR.body(fname).unwrap();
-        let mut si = SIRInterpreter::new(&*body);
-        // The raw pointer `tio` and the reference it was created from do not alias since we won't
-        // be using the reference until the function `interpret` returns.
-        si.set_trace_inputs(tio);
-        unsafe {
-            si.interpret(body);
-        }
-    }
-
-    #[test]
-    fn simple() {
-        struct IO(u8, u8);
-        #[no_mangle]
-        fn simple(io: &mut IO) {
-            let a = 3;
-            io.1 = a;
-        }
-        let mut tio = IO(0, 0);
-        interp("simple", &mut tio as *mut _ as *mut u8);
-        assert_eq!(tio.1, 3);
-    }
-
-    #[test]
-    fn tuple() {
-        struct IO((u8, u8, u8));
-        #[no_mangle]
-        fn func_tuple(io: &mut IO) {
-            let a = io.0;
-            let b = a.2;
-            (io.0).1 = b;
-        }
-
-        let mut tio = IO((1, 2, 3));
-        interp("func_tuple", &mut tio as *mut _ as *mut u8);
-        assert_eq!(tio.0, (1, 3, 3));
-    }
-
-    #[test]
-    fn reference() {
-        struct IO(u8, u8);
-        #[no_mangle]
-        fn func_ref(io: &mut IO) {
-            let a = 5u8;
-            let b = &a;
-            io.1 = *b;
-        }
-
-        let mut tio = IO(5, 0);
-        interp("func_ref", &mut tio as *mut _ as *mut u8);
-        assert_eq!(tio.1, 5);
-    }
-
-    #[test]
-    fn tupleref() {
-        struct IO((u8, u8));
-        #[no_mangle]
-        fn func_tupleref(io: &mut IO) {
-            let a = io.0;
-            (io.0).1 = 5; // Make sure the line above copies.
-            let b = &a;
-            (io.0).0 = b.1;
-        }
-
-        let mut tio = IO((0, 3));
-        interp("func_tupleref", &mut tio as *mut _ as *mut u8);
-        assert_eq!(tio.0, (3, 5));
-    }
-
-    #[test]
-    fn doubleref() {
-        struct IO((u8, u8));
-        #[no_mangle]
-        fn func_doubleref(io: &mut IO) {
-            let a = &io.0;
-            (io.0).0 = a.1;
-        }
-
-        let mut tio = IO((0, 3));
-        interp("func_doubleref", &mut tio as *mut _ as *mut u8);
-        assert_eq!(tio.0, (3, 3));
-    }
-
-    #[test]
-    fn call() {
-        struct IO(u8, u8);
-
-        fn foo(i: u8) -> u8 {
-            i
-        }
-
-        #[no_mangle]
-        fn func_call(io: &mut IO) {
-            let a = foo(5);
-            io.0 = a;
-        }
-
-        let mut tio = IO(0, 0);
-        interp("func_call", &mut tio as *mut _ as *mut u8);
-        assert_eq!(tio.0, 5);
-    }
-}
