--- internal_ws/ykcompile/build.rs
@@ -1,5 +1,0 @@
-fn main() {
-    cc::Build::new()
-        .file("src/test_helpers.c")
-        .compile("ykcompile_test_helpers");
-}
