+++ internal_ws/yktrace/src/swt.rs
@@ -216,37 +216,3 @@ unsafe fn yk_swt_stop_tracing_impl(ret_trace_len: &mut usize) -> *mut SwtLoc {
 
     return ret_trace;
 }
-
-#[cfg(test)]
-#[cfg(tracermode = "sw")]
-mod tests {
-    use crate::{test_helpers, TracingKind};
-
-    const TRACING_KIND: TracingKind = TracingKind::SoftwareTracing;
-
-    #[test]
-    fn trace() {
-        test_helpers::trace(TRACING_KIND);
-    }
-
-    #[test]
-    fn trace_twice() {
-        test_helpers::trace_twice(TRACING_KIND);
-    }
-
-    #[test]
-    fn trace_concurrent() {
-        test_helpers::trace_concurrent(TRACING_KIND);
-    }
-
-    #[test]
-    #[should_panic]
-    fn oob_trace_index() {
-        test_helpers::oob_trace_index(TRACING_KIND);
-    }
-
-    #[test]
-    fn in_bounds_trace_indices() {
-        test_helpers::in_bounds_trace_indices(TRACING_KIND);
-    }
-}
