+++ internal_ws/yktrace/src/tir.rs
@@ -605,113 +605,3 @@ impl fmt::Display for TirOp {
         }
     }
 }
-
-#[cfg(test)]
-pub mod test_helpers {}
-
-#[cfg(test)]
-pub mod tests {
-    use super::TirTrace;
-    use crate::{sir::SIR, start_tracing, trace_debug, TracingKind};
-    use fm::FMBuilder;
-    use regex::Regex;
-    use test::black_box;
-
-    /// Fuzzy matches the textual TIR for the trace `tt` with the pattern `ptn`.
-    fn assert_tir(ptn: &str, tt: &TirTrace) {
-        let ptn_re = Regex::new(r"%.+?\b").unwrap(); // Names are words prefixed with `%`.
-        let text_re = Regex::new(r"\$?.+?\b").unwrap(); // Any word optionally prefixed with `$`.
-        let matcher = FMBuilder::new(ptn)
-            .unwrap()
-            .name_matcher(ptn_re, text_re)
-            .distinct_name_matching(true)
-            .build()
-            .unwrap();
-
-        let res = matcher.matches(&format!("{}", tt));
-        if let Err(e) = res {
-            panic!("{}", e);
-        }
-    }
-
-    #[test]
-    fn nonempty_tir_trace() {
-        #[inline(never)]
-        #[interp_step]
-        fn work(io: &mut IO) {
-            let mut res = 0;
-            while res < io.1 {
-                res += io.0;
-            }
-            io.2 = res
-        }
-
-        struct IO(usize, usize, usize);
-        let mut io = IO(3, 13, 0);
-        let tracer = start_tracing(TracingKind::default());
-        black_box(work(&mut io));
-        let sir_trace = tracer.stop_tracing().unwrap();
-        let tir_trace = TirTrace::new(&*SIR, &sir_trace).unwrap();
-        assert_eq!(io.2, 15);
-        assert!(tir_trace.len() > 0);
-    }
-
-    struct DebugTirIO(usize, usize);
-
-    #[inline(never)]
-    #[interp_step]
-    fn debug_tir_work(io: &mut DebugTirIO) {
-        match io.0 {
-            0 => {
-                trace_debug("Add 10");
-                io.1 += 10;
-            }
-            1 => {
-                trace_debug("Minus 2");
-                io.1 -= 2;
-            }
-            2 => {
-                trace_debug("Multiply 2");
-                io.1 *= 2;
-            }
-            _ => unreachable!()
-        }
-    }
-
-    #[test]
-    fn trace_debug_tir() {
-        let mut io = DebugTirIO(0, 0);
-        let tracer = start_tracing(TracingKind::default());
-        black_box(debug_tir_work(&mut io)); // +10
-        black_box(debug_tir_work(&mut io)); // +10
-        io.0 = 2;
-        black_box(debug_tir_work(&mut io)); // *2
-        io.0 = 1;
-        black_box(debug_tir_work(&mut io)); // -2
-        let sir_trace = tracer.stop_tracing().unwrap();
-        let tir_trace = TirTrace::new(&*SIR, &sir_trace).unwrap();
-        assert_eq!(io.1, 38);
-        assert_tir(
-            "...\n\
-            ops:\n\
-              ...
-              // Add 10
-              ...
-              ... + 10usize (checked)
-              ...
-              // Add 10
-              ...
-              ... + 10usize (checked)
-              ...
-              // Multiply 2
-              ...
-              ... * 2usize (checked)
-              ...
-              // Minus 2
-              ...
-              ... - 2usize (checked)
-              ...",
-            &tir_trace
-        );
-    }
-}
