+++ tests/Cargo.toml
@@ -1,9 +1,15 @@
 [package]
-name = "ykrt"
+name = "tests"
 version = "0.1.0"
 authors = ["The Yorick Developers"]
 edition = "2018"
-license = "Apache-2.0 OR MIT"
 
 [dependencies]
+fm = "0.2.0"
 ykshim_client = { path = "../ykshim_client" }
+libc = "0.2.82"
+regex = "1.4.3"
+ykrt = { path = "../ykrt" }
+
+[build-dependencies]
+cc = "1.0.66"
