+++ tests/src/lib.rs
@@ -1,0 +1,13 @@
+#![feature(test)]
+
+#[cfg(test)]
+mod helpers;
+
+#[cfg(test)]
+mod ykbh;
+
+#[cfg(test)]
+mod ykcompile;
+
+#[cfg(test)]
+mod yktrace;
